Removed unused files from original fork
