--- src/abb_robot_client/robotraconteur/__main__.py
@@ -1,3 +1,0 @@
-from .abb_robotraconteur_rws_driver import main
-
-main()
