--- test/rapid/EIO.cfg
@@ -1,27 +1,0 @@
-EIO:CFG_1.0:6:1::
-#
-INDUSTRIAL_NETWORK:
-
-      -Name "DeviceNet" -Label "DeviceNet Master/Slave Network" -Address "2"
-#
-DEVICENET_DEVICE:
-
-      -Name "d651" -VendorName "ABB Robotics" -ProductName "Combi I/O Device"\
-      -Label "DSQC 651 Combi I/O Device" -ProductCode 25 -DeviceType 100\
-      -ConnectionType "COS" -OutputSize 5 -InputSize 1
-#
-EIO_SIGNAL:
-
-      -Name "test_digital_io2" -SignalType "DO" -Device "d651" -DeviceMap "32"\
-      -Access "All"
-
-      -Name "test_analog_io2" -SignalType "AO" -Device "d651" -DeviceMap "0-15"\
-      -Access "All" -EncType "UNSIGNED" -MaxLog 10 -MaxPhys 10 -MaxBitVal 65000
-
-      -Name "test_digital_io1" -SignalType "DO" -Access "All"
-
-      -Name "test_analog_io1" -SignalType "AO" -Access "All"
-
-      -Name "mode" -SignalType "AO" -Access "All" -EncType "UNSIGNED"
-
-      -Name "stop_egm" -SignalType "DO" -Access "All"
